Rename Integrations -> Plugins; modernize plugin card UI
Rename: codebase already says "plugin" everywhere (plugins/ directory,
plugin.json, /v1/plugins/* endpoints, create-plato-plugin scaffolder),
but the admin nav and route still said "Integrations". That label also
doesn't fit Phase 2 plugins (e.g. teacher-comments) that don't connect
external services. Aligning the UI with the code reduces friction for
both admins and plugin authors.

Changes:
  - Nav label "Integrations" -> "Plugins"
  - Route /plato/integrations -> /plato/plugins
  - Component AdminIntegrations -> AdminPlugins (file renamed; git -M
    keeps history)
  - /plato/integrations -> 302 redirect to /plato/plugins so any
    bookmarks keep working
  - All doc references updated (README, CLAUDE.md, docs/plugins/*,
    Slack plugin docs, SDK comments)

Card UI redesign — the previous design felt rough (per direct review):
the "Show settings" button rendered as plain text; the checkbox
toggle felt dated; capability chips dominated the header.

  - Switch component (WAI-ARIA role="switch", focusable, space-toggle)
    replaces the checkbox + label combination. Adds an "On"/"Off"
    indicator next to it for clarity.
  - Settings panel renders inline whenever the plugin is enabled
    (matches WordPress / Shopify Apps patterns) — disabled cards are
    subdued, enabled ones expose their controls inline so admins can
    verify state at a glance. The "Show settings" disclosure is gone.
  - Capability chips moved to a collapsed <details> "Show N permissions"
    affordance. Auditing them is one click; they no longer compete with
    the description for attention. Rendered as <code> tokens to make
    clear they're machine identifiers, not user-facing labels.
  - Disabled plugin cards get opacity-75 so the active set is visually
    distinct.
  - Page subtitle terser: "Enable, disable, and configure plato
    plugins. Build your own."

Server: 195 tests pass. Client: 45 tests pass. Lint clean. Build clean.
